moved sockets shit to another file, may have somehow broken music parsing as a result
